Removed 16 network limit. Prevent "Run Schedule" from being checked when networks are not connected. Fixed some compiler warnings. More changes to xTester, but RGB tests still NOT finished.
git-svn-id: svn://svn.code.sf.net/p/xlights/code/trunk@73 ad4834c4-2333-46e5-bf01-adba983f9b38

+++ xConverter/xConverter.h
@@ -85,13 +85,13 @@ class xConverter: public wxFrame
         wxUint8 * SeqData;
         long SeqDataLen;
         long SeqNumPeriods;
-        int SeqNumChannels;
-        int TotChannels;
+        size_t SeqNumChannels;
+        size_t TotChannels;
         wxArrayString FileNames;
         wxArrayString ChannelNames;
         wxArrayInt ChannelColors;
         std::vector<int> VixChannels;
-        std::vector<int> NetMaxChannel;
+        std::vector<size_t> NetMaxChannel;
 
 		DECLARE_EVENT_TABLE()
 };
